refactor : NewLevellogQnAListProps 위치 NewLevellogQnAList 컴포넌트 상단으로 이동
